requirements.txt
torch
torchvision
fastapi
uvicorn
jinja2
pillow
python-multipart
